backend/requirements.txt
fastapi
uvicorn
fastf1
pandas
openai
python-dotenv

